requirements.txt
fastapi==0.111.0
python-socketio==5.11.3
python-dotenv==1.0.1
redis==5.0.6